+++ src/string_utils.rs
@@ -1,0 +1,22 @@
+pub fn rem_first_and_last(value: &str) -> &str {
+    let mut chars = value.chars();
+    chars.next();
+    chars.next_back();
+    chars.as_str()
+}
+
+pub fn trim_whites(str: &str) -> String {
+    let s = str.trim();
+
+    let s: String = s.replace('\n', "");
+
+    s
+}
+
+pub fn trim_parentheses(str: &str) -> String {
+    let s = str.trim();
+
+    let s: String = s.replace(['(', ')'], "");
+
+    s
+}
